Extract submission validation to a concern
